[#116] refactor(v1_weather): move the CHIRPS normals builder into a management command
source/ holds data — topojson, CSVs, the normals rasters — so a runnable Python
script did not belong there. It is now a management command alongside
extract_weather_normals, which is what you run straight after it:

  python manage.py build_chirps_normals        # rebuild the raster (~1.6 GB in)
  python manage.py extract_weather_normals     # load it into the DB

Follows the self-contained shape of v1_iks download_iks_data (module helpers +
Command), and drops the hardcoded /app/source/30years/... output path in favour
of utils.raster_path(), so the filename has one owner in constants.NORMALS_RASTERS
instead of being repeated in the builder.

No behaviour change: same source, bbox, years, aggregation and band tags — only
prints become self.stdout, and a failed fetch raises CommandError rather than
printing to stderr.

Docs updated to the new invocation: the 30years README (now noting the directory
is data-only), the WX-5 design doc, and its work plan — whose N2 row still said
`normals.py`, a module that became utils.py during implementation.
